Delete an outdated-looking importer (can't find on GUI)
This importer looks to belong to "uploads" module, as it has function
names like uploads_admin_importgallery(). Also it contains unsafe SQL
command building, e.g. adding $filename with string concatenation. And
the Album classes are outdated with create_function() call. I simply
cannot find a GUI for this and cannot find a way to invoke this or any
reference to this - so deleting for now. (To stay on safe side.)